Modify to use activity instead of fragment
