--- Food_Tracker_App/app/src/main/res/layout/activity_add_recipe.xml
@@ -1,131 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".ui.recipes.AddRecipeActivity">
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="151dp"
-        android:layout_height="46dp"
-        android:text="Add a Recipe"
-        android:textSize="20sp" />
-
-    <EditText
-        android:id="@+id/recipeTitle"
-        android:layout_width="275dp"
-        android:layout_height="48dp"
-        android:layout_marginLeft = "40dp"
-        android:layout_marginTop = "10dp"
-        android:background="@color/background_dark"
-        android:ems="10"
-        android:hint="title"
-        android:inputType="textPersonName"
-        android:paddingLeft="20dp" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <EditText
-            android:id="@+id/recipePrepTime"
-            android:layout_width="133dp"
-            android:layout_height="48dp"
-            android:background="@color/background_dark"
-            android:ems="10"
-            android:hint="prep time"
-            android:inputType="number"
-            android:layout_marginLeft = "40dp"
-            android:layout_marginTop = "10dp"
-            android:paddingLeft="20dp" />
-
-        <EditText
-            android:id="@+id/recipeServings"
-            android:layout_width="127dp"
-            android:layout_height="48dp"
-            android:background="@color/background_dark"
-            android:ems="10"
-            android:hint="servings"
-            android:inputType="number"
-            android:layout_marginLeft = "20dp"
-            android:layout_marginTop = "10dp"
-            android:paddingLeft="20dp" />
-    </LinearLayout>
-
-    <EditText
-        android:id="@+id/recipeCategory"
-        android:layout_width="275dp"
-        android:layout_height="48dp"
-        android:background="@color/background_dark"
-        android:ems="10"
-        android:hint="catogry"
-        android:inputType="text"
-        android:layout_marginLeft = "40dp"
-        android:layout_marginTop = "10dp"
-        android:paddingLeft="20dp" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginLeft = "40dp"
-        >
-
-        <TextView
-            android:id="@+id/textView3"
-            android:layout_width="125dp"
-            android:layout_height="40dp"
-            android:text="Ingredients: " />
-
-        <Button
-            android:id="@+id/addIngredient_Button"
-            android:layout_width="75dp"
-            android:layout_height="48dp"
-
-            android:text="add" />
-    </LinearLayout>
-
-    <ListView
-        android:id="@+id/ingredients"
-        android:layout_width="175dp"
-        android:layout_height="125dp"
-        android:layout_marginTop = "10dp"
-        android:layout_marginLeft = "40dp"
-        />
-
-    <EditText
-        android:id="@+id/recipeComments"
-        android:layout_width="274dp"
-        android:layout_height="48dp"
-        android:ems="10"
-        android:hint="Comments"
-        android:layout_marginTop = "10dp"
-        android:layout_marginLeft = "40dp"
-        android:inputType="text" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="48dp"
-        android:orientation="horizontal"
-
-        android:layout_marginLeft = "40dp">
-
-        <Button
-            android:id="@+id/addIngredientConfirm"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Confirm" />
-
-        <Button
-            android:id="@+id/cancelAddIngredient"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft = "40dp"
-            android:text="Cancel" />
-    </LinearLayout>
-
-</LinearLayout>
